Move useCountdown.js to hooks directory and change the `handle` parameter to be optional
